requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.12.0
python-magic==0.4.27
supabase==2.13.0
requests==2.31.0
ffmpeg-python==0.2.0
jinja2==3.1.2
pydantic==2.4.2 